Move all combination list related components into a dedicated component folder
