+++ admin-dev/themes/new-theme/js/pages/product/components/combination-list/bulk-choices-selector.ts
