+++ admin-dev/themes/new-theme/js/pages/product/components/combination-list/combinations-list-editor.ts
@@ -25,7 +25,7 @@
 
 import ProductMap from '@pages/product/product-map';
 import ProductEventMap from '@pages/product/product-event-map';
-import CombinationsListRenderer from '@pages/product/combination/combinations-list-renderer';
+import CombinationsListRenderer from '@pages/product/components/combination-list/combinations-list-renderer';
 import {EventEmitter} from 'events';
 import {isUndefined} from '@PSTypes/typeguard';
 import BigNumber from '@node_modules/bignumber.js';
