+++ admin-dev/themes/new-theme/js/pages/product/components/combination-list/combinations-list-renderer.ts
